refactored the way we take curriculum and fr_curve.outsteps from values to percentage

+++ src/qubit/dataclasses/training_config.py
@@ -11,7 +11,7 @@ class TrainingConfig:
     prediction_mode: PredictionMode
     epochs: int 
     batch_size: int
-    curriculum: List[int]
+    curriculum: List[float] # List[int]
     verbose : VerboseMode
     phases: List[PhaseConfig]
     fr_eval: FrEvalConfig
